requirements.txt
flask
onnxruntime
numpy
pillow
facenet-pytorch
gdown
